Get rid of some valid compiler warnings with CodeWarrior that were reported by Paul Chambers.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@2845 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Browser.cxx
@@ -1,5 +1,5 @@
 //
-// "$Id: Fl_Browser.cxx,v 1.9.2.12.2.6 2002/08/09 01:09:48 easysw Exp $"
+// "$Id: Fl_Browser.cxx,v 1.9.2.12.2.7 2002/11/19 16:37:34 easysw Exp $"
 //
 // Browser widget for the Fast Light Tool Kit (FLTK).
 //
@@ -164,7 +164,7 @@ void Fl_Browser::insert(int line, FL_BLINE* t) {
 void Fl_Browser::insert(int line, const char* newtext, void* d) {
   int l = strlen(newtext);
   FL_BLINE* t = (FL_BLINE*)malloc(sizeof(FL_BLINE)+l);
-  t->length = l;
+  t->length = (short)l;
   t->flags = 0;
   strcpy(t->txt, newtext);
   t->data = d;
@@ -185,7 +185,7 @@ void Fl_Browser::text(int line, const char* newtext) {
     replacing(t, n);
     cache = n;
     n->data = t->data;
-    n->length = l;
+    n->length = (short)l;
     n->flags = t->flags;
     n->prev = t->prev;
     if (n->prev) n->prev->next = n; else first = n;
@@ -495,5 +495,5 @@ int Fl_Browser::value() const {
 }
 
 //
-// End of "$Id: Fl_Browser.cxx,v 1.9.2.12.2.6 2002/08/09 01:09:48 easysw Exp $".
+// End of "$Id: Fl_Browser.cxx,v 1.9.2.12.2.7 2002/11/19 16:37:34 easysw Exp $".
 //
